pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=0.3.23",
    "fastapi<0.100.0",
    "httpx>=0.28.1",
    "pydantic<2.0.0",
    "pypdf>=6.6.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "redis>=7.1.0",
    "sentence-transformers>=5.2.0",
    "uvicorn>=0.40.0",
    "faster-whisper>=1.1.0",
    "edge-tts>=6.1.12",
    "sqlalchemy[asyncio]>=2.0.46",
    "asyncpg>=0.31.0",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.5.0",
    "slowapi>=0.1.9",
    "email-validator>=2.3.0",
    "bcrypt==3.2.2",
]
